pyproject.toml
[project]
name = "ballsy-voice-assistant"
version = "1.0.0"
description = "AI Voice Assistant with Siri-like UI"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0", 
    "flask>=2.3.0",
    "python-dotenv>=1.0.0",
    "google-genai>=0.2.0",
    "SpeechRecognition>=3.10.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
    "websockets>=11.0.0",
    "jinja2>=3.1.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.12.0"
]
